+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes between Rockpool releases will be documented in this file.
 * Initial module structure for Xylo-IMU support
   - ``XyloIMUMonitor`` implemented
   - IMU preprocessing interface
-    - `Quantizer` , `RotationLookUpTable` , `JSVD` , `SubSpace` , `SampleAndHold` , `RotationRemoval` are implemented
+    - Sub-modules : `Quantizer` , `RotationLookUpTable` , `JSVD` , `SubSpace` , `SampleAndHold` , `RotationRemoval` , `BandPassFilter` , `FilterBank` , `HalfWaveRectifier` , `FullWaveRectifier` , `ScaleSpikeEncoder` , `IAFSpikeEncoder` are implemented
+    - The high-level simulation module encapsulating sub-modules `IMUIFSim` implemented.
 
 ### Changed